requirements.txt
ultralytics==8.0.0
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
matplotlib>=3.3.0
numpy>=1.19.0
albumentations>=1.0.0
tqdm>=4.60.0
pandas>=1.1.0
PyYAML>=5.4.0